LocalBackend: avoid setTimeout
It is necessary to delay parsing the textContent of a local backend
until after the DOM has been fully constructed. Using setTimeout for
doing that is unreliable and leads to random failures in the test suite.
This is normally not an issue in reality but the test suite relies on
data being available soon after the window `load` event has triggered.