Refactor backend APIs
This commit changes the APIs used by backend internally. It should not
(or only minimally) change the behavior of public APIs. The main change
is that

* Set operations are optionally asynchronous and can return promises.
  If supported by the backend, this allows BackendValue.set() to
  asynchronously succeed or fail.
* Backends now may emit information about paths. This allows
  BackendValues to know whether or not a paths can be modified or
  subscribed to.
* BackendValues cannot be modified anymore while they are not connected
  to a backend. Doing so will result in BackendValue.set() throwing an
  exception. In order to regain the 'old' behavior, a proxy backend must
  be used which caches set calls until the backend is actually connected
  to.

This commit also comes with more tests for different backend types and a
NodeJS compatible index file which can be used to implement backends in
nodejs.