TemplateComponents: better handling of .property$
Previously, for each template dependency a dynamic value was created
implicitly. References inside of a template to properties of the form
this.<NAME>$ would be interpreted to reference the property <NAME>.
This implicit handling has been removed, the property name is taken
as-is. The <NAME>$ getters are still available if no dependency with
the same name exists.