Backends: Introduce a special value to signal deleted values
In dynamic backends in can happen that properties which have previously
been there are removed. In order to signal that between backends and
bindings we introduce a new special value AWML.DELETED which is used to
signal that. Along with that we introduce a method clearValue() on
backends which is used to trigger that delete.