Backends.aes70: Signal children as deleted if they disappear
If objects are removed from a block we now signal them as deleted to the
corresponding binding. This then recursively also holds for all
properties in the subtree. This solves situations where otherwise a
binding might still contain the last value while we already know that it
does not exist, anymore.